requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1
pytest==7.4.3
pytest-mock==3.12.0

